fix: Screen height and builds warnings
